requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
alembic==1.7.5
SQLAlchemy==1.4.27
passlib==1.7.4
python_jose==3.3.0
psycopg2-binary==2.9.0
bcrypt==3.2.0